Move db setup code to plugins file
